feat: 添加 LazyLoad 功能的常量和插件实现
