--- packages/vue3/source/rspack/plugins/lcap.js
@@ -1,418 +1,0 @@
-const rspack = require('@rspack/core');
-
-const plugin = 'lcap-plugin';
-
-module.exports = class LcapPlugin {
-  constructor(options) {
-    this.options = options;
-  }
-  apply(compiler) {
-    const { isDev, isIncremental, lastResource, extra } = this.options;
-
-    // 增量发布
-    if (isIncremental) {
-      addIncrementalBuild(compiler, {
-        lastResource,
-      });
-
-      return;
-    }
-
-    // 标准发布
-    overrideRuntime(compiler, {
-      isDev,
-    });
-    emitChunksMapResource(compiler, {
-      isDev,
-    });
-    emitClientResource(compiler, {
-      isDev,
-      extra,
-    });
-  }
-}
-
-// 复写加载css\js chunk的函数
-function overrideRuntime(compiler, options) {
-  const { isDev } = options;
-
-  if (!isDev) {
-    return;
-  }
-
-  const { RuntimeModule, RuntimeGlobals } = compiler.webpack;
-
-  // 复写__webpack_require__.k和__webpack_require__.u
-  class OverrideRuntimeModule extends RuntimeModule {
-    constructor() {
-      super('override');
-    }
-
-    generate() {
-      return `// override
-      __webpack_require__.k = (chunkId) => {
-        return "" + window.lcapChunksNameMap[chunkId] + "." + window.lcapChunksHashMap[chunkId] + ".css"
-      }
-      // override
-      __webpack_require__.u = (chunkId) => {
-        return "" + window.lcapChunksNameMap[chunkId] + "." + window.lcapChunksHashMap[chunkId] + ".js"
-      }`;
-    }
-  }
-
-  compiler.hooks.thisCompilation.tap(plugin, (compilation) => {
-    compilation.hooks.runtimeRequirementInTree
-      .for(RuntimeGlobals.ensureChunkHandlers)
-      .tap(plugin, (chunk) => {
-        if (chunk.name === 'runtime') {
-          compilation.addRuntimeModule(chunk, new OverrideRuntimeModule());
-        }
-      });
-  });
-}
-
-// 增加chunkMap资源
-function emitChunksMapResource(compiler, options) {
-  const { isDev } = options;
-
-  if (!isDev) {
-    return;
-  }
-
-  compiler.hooks.emit.tapAsync(plugin, (compilation, callback) => {
-    // 2、增加router.min.js asset
-    const ChunksNameMap = {};
-    const ChunksHashMap = {};
-    const statsJson = compilation.getStats().toJson({
-      warnings: false,
-      modules: false,
-    });
-
-    statsJson.chunks.forEach(chunk => {
-      const chunkId = chunk.id;
-      const chunkName = chunk.names[0];
-      const chunkHash = chunk.hash.substring(0, 8);
-
-      ChunksNameMap[chunkId] = chunkName || chunkId;
-      ChunksHashMap[chunkId] = chunkHash;
-    });
-
-    const { RawSource } = compiler.webpack.sources;
-
-    const code = `window.lcapChunksNameMap = ${JSON.stringify(ChunksNameMap)};
-window.lcapChunksHashMap = ${JSON.stringify(ChunksHashMap)};
-
-var lcap_changed_chunks = [];
-
-// lcap_changed_chunks placeholder
-
-lcap_changed_chunks.forEach(item => {
-  window.lcapChunksNameMap[item.id] = item.name;
-  window.lcapChunksHashMap[item.id] = item.hash;
-});`;
-
-    compilation.emitAsset('router.min.js', new RawSource(Buffer.from(code)), { minimized: true })
-
-    callback();
-  })
-}
-
-// 生成client.js
-function emitClientResource(compiler, options) {
-  const { isDev, extra } = options;
-  compiler.hooks.compilation.tap(plugin, compilation => {
-      rspack.HtmlRspackPlugin.getCompilationHooks(compilation).beforeAssetTagGeneration.tapPromise(plugin, async data => {
-        const { assets } = data;
-        const { publicPath, js, css } = assets;
-
-        const allJS = [
-          ...(extra.preJS || []),
-          ...js,
-          ...(extra.postJS || []),
-        ];
-
-        const allCSS = [
-          ...(extra.preCSS || []),
-          ...css,
-          ...(extra.postCSS || []),
-        ];
-
-        const clientCode = `(function() {
-          var LazyLoad = (function (doc) {
-            var env,
-                head,
-                pending = {},
-                pollCount = 0,
-                queue = {
-                  css: [],
-                  js: []
-                },
-                styleSheets = doc.styleSheets;
-
-            function createNode(name, attrs) {
-              var node = doc.createElement(name),
-                      attr;
-              for (attr in attrs) {
-                if (attrs.hasOwnProperty(attr)) {
-                  node.setAttribute(attr, attrs[attr]);
-                }
-              }
-              return node;
-            }
-
-            function finish(type) {
-              var p = pending[type],
-                      callback,
-                      urls;
-              if (p) {
-                callback = p.callback;
-                urls = p.urls;
-                urls.shift();
-                pollCount = 0;
-                if (!urls.length) {
-                  callback && callback.call(p.context, p.obj);
-                  pending[type] = null;
-                  queue[type].length && load(type);
-                }
-              }
-            }
-
-            function getEnv() {
-              var ua = navigator.userAgent;
-              env = {
-                async: doc.createElement('script').async === true
-              };
-              (env.webkit = /AppleWebKit\//.test(ua)) || (env.ie = /MSIE|Trident/.test(ua)) || (env.opera = /Opera/.test(ua)) || (env.gecko = /Gecko\//.test(ua)) || (env.unknown = true);
-            }
-
-            function load(type, urls, callback, obj, context) {
-              var _finish = function () {
-                        finish(type);
-                      },
-                      isCSS = type === 'css',
-                      nodes = [],
-                      i,
-                      len,
-                      node,
-                      p,
-                      pendingUrls,
-                      url;
-              env || getEnv();
-              if (urls) {
-                urls = typeof urls === 'string' ? [urls] : urls.concat();
-                if (isCSS || env.async || env.gecko || env.opera) {
-                  queue[type].push({
-                    urls: urls,
-                    callback: callback,
-                    obj: obj,
-                    context: context
-                  });
-                } else {
-                  for (i = 0, len = urls.length; i < len; ++i) {
-                    queue[type].push({
-                      urls: [urls[i]],
-                      callback: i === len - 1 ? callback : null,
-                      obj: obj,
-                      context: context
-                    });
-                  }
-                }
-              }
-              if (pending[type] || !(p = pending[type] = queue[type].shift())) {
-                return;
-              }
-              head || (head = doc.head || doc.getElementsByTagName('head')[0]);
-              pendingUrls = p.urls.concat();
-              for (i = 0, len = pendingUrls.length; i < len; ++i) {
-                url = pendingUrls[i];
-                if (isCSS) {
-                  node = env.gecko
-                          ? createNode('style')
-                          : createNode('link', {
-                            href: url,
-                            rel: 'stylesheet'
-                          });
-                } else {
-                  node = createNode('script', {
-                    src: url
-                  });
-                  node.async = false;
-                }
-                node.className = 'lazyload';
-                node.setAttribute('charset', 'utf-8');
-                if (env.ie && !isCSS && 'onreadystatechange' in node && !('draggable' in node)) {
-                  node.onreadystatechange = function () {
-                    if (/loaded|complete/.test(node.readyState)) {
-                      node.onreadystatechange = null;
-                      _finish();
-                    }
-                  };
-                } else if (isCSS && (env.gecko || env.webkit)) {
-                  if (env.webkit) {
-                    p.urls[i] = node.href;
-                    pollWebKit();
-                  } else {
-                    node.innerHTML = '@import "' + url + '";';
-                    pollGecko(node);
-                  }
-                } else {
-                  node.onload = node.onerror = _finish;
-                }
-                nodes.push(node);
-              }
-              for (i = 0, len = nodes.length; i < len; ++i) {
-                head.appendChild(nodes[i]);
-              }
-            }
-
-            function pollGecko(node) {
-              var hasRules;
-              try {
-                hasRules = !!node.sheet.cssRules;
-              } catch (ex) {
-                pollCount += 1;
-                if (pollCount < 200) {
-                  setTimeout(function () {
-                    pollGecko(node);
-                  }, 50);
-                } else {
-                  hasRules && finish('css');
-                }
-                return;
-              }
-              finish('css');
-            }
-
-            function pollWebKit() {
-              var css = pending.css,
-                      i;
-              if (css) {
-                i = styleSheets.length;
-                while (--i >= 0) {
-                  if (styleSheets[i].href === css.urls[0]) {
-                    finish('css');
-                    break;
-                  }
-                }
-                pollCount += 1;
-                if (css) {
-                  if (pollCount < 200) {
-                    setTimeout(pollWebKit, 50);
-                  } else {
-                    finish('css');
-                  }
-                }
-              }
-            }
-            function preload(url) {
-              var asAttr = '';
-              if (/\.js$/.test(url)) {
-                asAttr = 'script';
-              } else if (/\.css$/.test(url)) {
-                asAttr = 'style';
-              }
-              var node = createNode('link', {
-                href: url,
-                as: asAttr,
-                rel: 'preload'
-              });
-              head.appendChild(node);
-            }
-            return {
-              css: function (urls, callback, obj, context) {
-                load('css', urls, callback, obj, context);
-              },
-              js: function (urls, callback, obj, context) {
-                load('js', urls, callback, obj, context);
-              },
-              preload: function (urls) {
-                if (Object.prototype.toString.call(urls) === '[object Array]') {
-                  for (var i = 0; i < urls.length; i++) {
-                    urls[i] && preload(urls[i]);
-                  }
-                } else {
-                  urls && preload(urls);
-                }
-              },
-              init: function () {
-                pending = {
-                  css: [],
-                  js: []
-                };
-                queue = {
-                  css: [],
-                  js: []
-                };
-              }
-            };
-          })(document);
-          function loadAssets() {
-            // chunksMap
-            ${isDev} && LazyLoad.js(["${publicPath}router.min.js"].map(item => {
-              var timestamp = Date.now();
-              return item + '?timestamp=' + timestamp;
-            }));
-
-            LazyLoad.js(${JSON.stringify(allJS)});
-            LazyLoad.css(${JSON.stringify(allCSS)});
-          }
-
-          ${extra.entryCode}
-        })();
-        `
-        compilation.emitAsset('client.js', new compiler.webpack.sources.RawSource(Buffer.from(clientCode)), { minimized: true });
-
-        data.assets.js = [`${publicPath}client.js?t=${Date.now()}`];
-        data.assets.css = [];
-      });
-    });
-}
-
-// 增量构建
-function addIncrementalBuild(compiler, options) {
-  const { lastResource } = options;
-  const { chunksMap: chunksMapCode } = lastResource;
-
-  compiler.hooks.emit.tapAsync(plugin, (compilation, callback) => {
-    const statsJson = compilation.getStats().toJson({
-      warnings: false,
-      modules: false,
-    });
-
-    // 只保留pages相关的chunk 资源
-    const pagesAssets = [];
-    const changedChunks = [];
-
-    statsJson.chunks.filter(chunk => {
-      if (chunk.idHints?.includes('page')) {
-        pagesAssets.push(...chunk.files);
-        changedChunks.push({
-          id: chunk.id,
-          name: chunk.names[0] || chunk.id,
-          hash: chunk.hash.substring(0, 8),
-        });
-      }
-    });
-
-    // 删除不需要的chunk(非页面相关的chunk)
-    statsJson.assets.forEach(asset => {
-      if (!pagesAssets.includes(asset.name)) {
-        compilation.deleteAsset(asset.name);
-      }
-    })
-
-    // 生成增量变更的chunks-map.js
-    let code = chunksMapCode;
-
-    const placeholderIdx = code.indexOf('// lcap_changed_chunks placeholder');
-    // 前面插入代码
-    if (placeholderIdx !== -1) {
-      code = code.slice(0, placeholderIdx) + `lcap_changed_chunks.push(...${JSON.stringify(changedChunks)});\n` + code.slice(placeholderIdx);
-    }
-
-    const { RawSource } = compiler.webpack.sources;
-    compilation.emitAsset('router.min.js', new RawSource(Buffer.from(code)), { minimized: true })
-
-    callback();
-  });
-}
